+++ .eslintrc.js
@@ -10,7 +10,8 @@ module.exports = {
     },
     env: {
         es6: true,
-        node: true
+        node: true,
+        mocha: true
     },
     extends: 'eslint:recommended'
 };